+++ src/templates/home/styles.ts
@@ -1,0 +1,5 @@
+import { css } from "@emotion/react";
+
+export const lottieStyles = css`
+	margin: auto;
+`;
